+++ public/javascripts/reportTrafficLight.js
@@ -8,30 +8,20 @@ const myconfig = config.get('TrafficLight.lightManager');
 var currentLight = 0;
 var debug = !myconfig.lightsEnabled;
 
-exports.initLight = function() {
+exports.initReport = function() {
     logger.debug('Init Light (DebugMode: '+debug+')');
     setOn(1,1,1); //green, yellow, red
 }
 
-exports.setLight = function(light) {
-    //logger.debug('Check '+ currentLight +' =='+ light);
-    if (currentLight != light) {
-        logger.debug('Switch light to '+ light);
-        light==1?1:0
-        setOn(light==1?1:0, light==2?1:0,light==3?1:0)
+exports.reportStatusChange = function(changedAlarm, oldStatus, newStatus, alertList) {
+    if (oldStatus.value != newStatus.value) {
+        light = newStatus.value;
+        if (currentLight != light) {
+            light == 1 ? 1 : 0
+            setOn(light == 1 ? 1 : 0, light == 2 ? 1 : 0, light == 3 ? 1 : 0)
+        }
     }
 }
-
-/*
-function onOrOffValue(sollColor, ledNumber) {
-    if (sollColor == ledNumber) {
-        currentLight = sollColor;
-        return 1;
-    } else {
-        return 0;
-    }
-}
-*/
 
 function setOn(green, yellow, red) {
     if (!debug) runner.runTrafficLight(green, yellow, red);