+++ CHANGELOG.md
@@ -9,9 +9,15 @@
 
 
 ## Changelog
+### 1.1.0
+* Count of checks is displayed
+* Better group namse for DTSMon based on hostname 
+* New Report Plugin for insert a measurement into influx
+
 ### 1.0.0
 * Some code cleanup
 * Introduce a alarm delay for DTSMon checks
+
 ### 0.9.3
 * Added state related favicon
 * Autoreload of the statuspage an visualisation of last reload timestamp